Plot Wavelet Variance

This function displays a plot of wavelet variance accounting for confidence interval values and supplied efficiency. This function is a slight modification of the plot.wvar function provided in the wv package.

plotwvar(
  x,
  units = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col_wv = NULL,
  col_ci = NULL,
  nb_ticks_x = NULL,
  nb_ticks_y = NULL,
  legend_position = NULL,
  ci_wv = NULL,
  point_cex = NULL,
  point_pch = NULL
)

Arguments

x

A wvar object.

units

A string that specifies the units of time plotted on the x axis.

xlab

A string that gives a title for the x axis.

ylab

A string that gives a title for the y axis.

main

A string that gives an overall title for the plot.

col_wv

A string that specifies the color of the wavelet variance line.

col_ci

A string that specifies the color of the confidence interval polygon.

nb_ticks_x

An integer that specifies the maximum number of ticks for the x-axis.

nb_ticks_y

An integer that specifies the maximum number of ticks for the y-axis.

legend_position

A string that specifies the position of the legend (use legend_position = NA to remove legend).

ci_wv

A boolean that determines whether a confidence interval polygon will be drawn.

point_cex

A double that specifies the size of each symbol to be plotted.

point_pch

A double that specifies the symbol type to be plotted.

...

Additional arguments affecting the plot.

Value

Plot of wavelet variance and confidence interval for each scale.

Author

Stephane Guerrier